changed module and DTO handling
The program now checks weither uio_pruss module is loaded, if it is loaded it will remove it, if not, it will do nothing.
The device three overlay EBB-PRU-ADC is not loaded by the program anymore. It should be added in /boot/uEnv.txt and in /etc/default/capemgr so it will be loaded on boot.